requirements.txt
asn1ate==0.6.0
kerberos==1.3.0
passlib==1.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyparsing==2.4.0
python-pam==1.8.4
